requirements.txt
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.0
joblib==1.5.2
kiwisolver==1.4.9
matplotlib==3.10.7
numpy==2.3.5
packaging==25.0
paho-mqtt==2.1.0
pandas==2.3.3
pillow==12.0.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.2
scipy==1.16.3
six==1.17.0
threadpoolctl==3.6.0
tzdata==2025.2
